Rename alquimia_ai to alquimia module integration
