+++ integrations/catalog.json
@@ -4,7 +4,7 @@
   "catalog_url": "https://raw.githubusercontent.com/github/spec-kit/main/integrations/catalog.json",
   "integrations": {
     "alquimia": {
-      "id": "alquimia",
+      "id": "alquimia_ai",
       "name": "Alquimia AI",
       "version": "1.0.0",
       "description": "Alquimia AI CLI integration",
